Use enum and use pull hierarchy creationg out into class w/ toplevel static method
