+++ app/src/org/odk/collect/android/logic/HierarchyEntryType.java
@@ -1,0 +1,8 @@
+package org.odk.collect.android.logic;
+
+/**
+ * @author Phillip Mates (pmates@dimagi.com).
+ */
+public enum HierarchyEntryType {
+    child, expanded, collapsed, question
+}
